+++ CHANGELOG.md
@@ -3,6 +3,28 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [1.5.0](https://github.com/cornerstonejs/cornerstone3D/compare/v1.4.6...v1.5.0) (2023-07-18)
+
+### Features
+
+- **viewportStatus:** Have renderedState to store the status of whether an image has been rendered yet ([#694](https://github.com/cornerstonejs/cornerstone3D/issues/694)) ([eeef233](https://github.com/cornerstonejs/cornerstone3D/commit/eeef233ef6e0b31f5b9ad14c6c722f529836ef5d))
+
+## [1.4.6](https://github.com/cornerstonejs/cornerstone3D/compare/v1.4.5...v1.4.6) (2023-07-14)
+
+**Note:** Version bump only for package root
+
+## [1.4.5](https://github.com/cornerstonejs/cornerstone3D/compare/v1.4.4...v1.4.5) (2023-07-14)
+
+### Bug Fixes
+
+- **color volume:** take into account number of components for the volume length ([#687](https://github.com/cornerstonejs/cornerstone3D/issues/687)) ([667c42e](https://github.com/cornerstonejs/cornerstone3D/commit/667c42e635f6262225f88aa458fe7482ea9ecf1e))
+
+## [1.4.4](https://github.com/cornerstonejs/cornerstone3D/compare/v1.4.3...v1.4.4) (2023-07-13)
+
+### Bug Fixes
+
+- **PT measurement units:** Non-SUV scaled, but pre-scaled PTs should show proper units ([#686](https://github.com/cornerstonejs/cornerstone3D/issues/686)) ([e9190df](https://github.com/cornerstonejs/cornerstone3D/commit/e9190df44b29be504b46ebb768e6ad2e6b02bbe3))
+
 ## [1.4.3](https://github.com/cornerstonejs/cornerstone3D/compare/v1.4.2...v1.4.3) (2023-07-12)
 
 ### Bug Fixes